+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @navikt/sif-common-ui
 
+## 0.0.19
+
+### Patch Changes
+
+- @navikt/sif-common-core-ds@10.2.19
+- @navikt/sif-common-formik-ds@2.4.16
+- @navikt/sif-common-utils@3.49.9
+- @navikt/sif-validation@1.0.2
+
+## 0.0.18
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/sif-common-formik-ds@2.4.16
+    - @navikt/sif-common-core-ds@10.2.18
+    - @navikt/sif-common-utils@3.49.9
+    - @navikt/sif-validation@1.0.2
+
 ## 0.0.17
 
 ### Patch Changes